requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
openai==1.12.0
anthropic==0.18.1
langchain==0.1.9
tiktoken==0.6.0
transformers==4.38.2
pandas>=2.2.0
numpy==1.26.4
scikit-learn==1.4.1.post1
gunicorn==21.2.0
pytest>=8.0.0
pytest-cov==4.1.0
black==24.2.0
isort==5.13.2
flake8==7.0.0
jinja2==3.1.3
WTForms==3.1.2
flask-wtf==1.2.1
tqdm>=4.66.1
requests==2.31.0
pyyaml==6.0.1
spacy==3.7.4
streamlit>=1.31.0
plotly>=5.18.0
altair==5.2.0
pillow==10.2.0
matplotlib==3.8.3 